Adicionado banco de dados da Amazon

+++ src/main/java/br/com/syntaxerror/syntaxerror/controller/ProfessorController.java
@@ -20,7 +20,7 @@ public class ProfessorController {
 
 
     @GetMapping("/professor/{codigo}")
-    public ProfessorEntity buscarProfessorPorMatricula(@PathVariable("codigo") BigInteger matricula){
+    public ProfessorEntity buscarProfessorPorMatricula(@PathVariable("codigo") Integer matricula){
         return professorService.buscarProfessorMatricula(matricula);
     }
