+++ src/main/java/br/com/syntaxerror/syntaxerror/repository/ProfessorRepository.java
@@ -4,8 +4,8 @@ import br.com.syntaxerror.syntaxerror.model.entity.ProfessorEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+
 
 @Repository
-public interface ProfessorRepository extends JpaRepository<ProfessorEntity, BigInteger> {
+public interface ProfessorRepository extends JpaRepository<ProfessorEntity, Integer> {
 }
